requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-storages==1.13.2
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
tomli==2.0.1
urllib3==1.26.13
